requirements-local.txt
python-dotenv
supabase
MetaTrader5
numpy
httpx
praw
requests
yfinance
schedule
selenium
